feat: added authentication and styling

+++ backend/src/services/posts.js
@@ -1,17 +1,13 @@
 import { Post } from '../db/models/post.js'
+import { User } from '../db/models/user.js'
 
-export async function createPost({
-  title,
-  author,
-  imgurl,
-  description,
-  ingredients,
-  directions,
-  tags,
-}) {
+export async function createPost(
+  userId,
+  { title, imgurl, description, ingredients, directions, tags },
+) {
   const post = new Post({
     title,
-    author,
+    author: userId,
     imgurl,
     description,
     ingredients,
@@ -30,9 +26,12 @@ async function listPosts(
 export async function listAllPosts(options) {
   return await listPosts({}, options)
 }
-export async function listPostsByAuthor(author, options) {
-  return await listPosts({ author }, options)
+export async function listPostsByAuthor(authorUsername, options) {
+  const user = await User.findOne({ username: authorUsername })
+  if (!user) return []
+  return await listPosts({ author: user._id }, options)
 }
+
 export async function listPostsByTag(tags, options) {
   return await listPosts({ tags }, options)
 }
@@ -41,15 +40,15 @@ export async function getPostById(postId) {
   return await Post.findById(postId)
 }
 export async function updatePost(
+  userId,
   postId,
-  { title, author, imgurl, description, ingredients, directions, tags },
+  { title, imgurl, description, ingredients, directions, tags },
 ) {
   return await Post.findOneAndUpdate(
-    { _id: postId },
+    { _id: postId, author: userId },
     {
       $set: {
         title,
-        author,
         imgurl,
         description,
         ingredients,
@@ -60,6 +59,6 @@ export async function updatePost(
     { new: true },
   )
 }
-export async function deletePost(postId) {
-  return await Post.deleteOne({ _id: postId })
+export async function deletePost(userId, postId) {
+  return await Post.deleteOne({ _id: postId, author: userId })
 }
